+++ src/set/isEmpty.lua
@@ -1,0 +1,5 @@
+local function isEmpty(value: { [any]: any }): boolean
+    return next(value) == nil
+end
+
+return isEmpty