+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- add `Set.count` and `Set.isEmpty` ([#21](https://github.com/seaofvoices/luau-disk/pull/21))
 - add `Set.map` ([#20](https://github.com/seaofvoices/luau-disk/pull/20))
 - add `Set.merge` ([#19](https://github.com/seaofvoices/luau-disk/pull/19))
 - add `Array.count` ([#18](https://github.com/seaofvoices/luau-disk/pull/18))